Refactor: Update event_date filter in EventRepositoryMongo
This commit updates the event_date filter in the EventRepositoryMongo class to include the start and end of the specified date. The start of the day is set to 00:00:00 and the end of the day is set to 23:59:59. This ensures that events within the specified date range are correctly retrieved.